Fix image URLs in exercise descriptions for A+
The A+ frontend must convert relative URLs in the course materials
to absolute since the URLs must refer to the backend server,
not to A+, when the material is viewed through A+.
<img> source URLs in exercise descriptions were not handled
correctly in the A+ frontend. They worked in chapter contents but
not in exercise descriptions because the service URLs for chapters
are based on the MOOC grader static URL while the exercise service
URLs are different than the static URL.

This fix adds data-aplus-path attribute to the img and similar
elements that link to a local file. This is done only in exercises,
not in the chapter contents. The attribute allows A+ to set
the image URL correctly.