Add the course head URLs setting to the index.yaml
The head URLs value is defined in the Sphinx conf.py with the
option course_head_urls (list of strings). If it is not set,
the course index.yaml or the exported JSON of the mooc-grader
do not define the head_urls key and the A+ course import does not
overwrite a potentially existing manually set value.