Remove the unused value from data-aplus-chapter
The quotes ("") in HTML attribute values cause issues in YAML
files. If a long YAML string is wrapped in double quotes,
then double queotes inside the string would have to be escaped.
However, YAML strings in single quotes may read escape characters
verbatim, which breaks HTML in A+. Furthermore, the code in
html_tools.py rewrites both HTML and YAML files, and in HTML files
the quotes around attributes must not be escaped.
Therefore, it is better to remove the quotes completely.

A+ frontend does not use the value of the HTML attribute
data-aplus-chapter when it first arrives from the exercise
service.