Add neutral options and partial points to pick-any (#47)
Add the `partial-points` option to the pick-any directive.
When set, partially correct answers award some points.
Over half of the checkboxes must be answered correctly in order
to gain some points. Previously, the question had to be answered
fully correctly in order to gain points.

Add neutral options that do not affect the grading in pick-any
questions. They are marked with `?` instead of `*` in RST.