Do not build plain titles if title|i18n is defined (#52)
Previously, if title|i18n was defined in config.yaml (used with
the submit directive), built files often contained
"title: None" as well. Now, plain title values are not
written into built files if title|i18n exists and hand-written
titles in the RST are saved into the title|i18n itself.

If config.language is not set correctly, the manually written
title is saved under the wrong language in title|i18n.
Because the yaml files are first built separately for different
languages and then merged together, the final title|i18n
could contain the original value from the config.yaml for
some language keys. This is not really an issue since
if somebody defines title|i18n by hand (not really recommended),
then it makes even less sense to also define language-specific
titles in RST (the title option of the submit directive).

If no title is defined in the the title option of the submit
directive nor the config.yaml file, a default title is formed
from the exercise key.

Co-authored-by: Markku Riekkinen <markku.riekkinen@aalto.fi>