Fix max points for 'pick randomly' questionnaires (#53)
The max points for a questionnaire with the pick randomly option
can not be equal to the sum of all question points, since
not all questions are used at the same time (a subset of questions
is randomly selected). This commit fixes how the max points
are automatically calculated from the questions or checked when
the max points value is also manually defined in the questionnaire
directive parameters.

We assume that all questions should have equal points when
'pick randomly' is used. A warning is printed if the question
points are not equal. The max points for the questionnaire
are then simply multiplied from the points of the first question.

Co-authored-by: Markku Riekkinen <markku.riekkinen@aalto.fi>