RST language detection may be toggled on or off
Detecting the language from RST files is not robust, so it is
necessary to sometimes disable it.
It works for a filename like "chapter1_fi.rst", but breaks for
a filename like "chapter_1b.rst".
This commit adds a new boolean setting for the course conf.py file
that enables or disables the language detection.

The language detection was added in commit
https://github.com/Aalto-LeTech/a-plus-rst-tools/commit/3ac1b3874d198c52c7c637ec2fa381fc1bbb6648