Refactor instructions in the submit directive body
The commit
https://github.com/Aalto-LeTech/a-plus-rst-tools/commit/3cdbd80a5991a902059f7d6df051d907e8236947
added support for writing the exercise description in the body of
the submit directive. However, the implementation was not clean.
The new implementation uses the API that already existed in the
a-plus-rst-tools code base and is intended exactly for this kind
of use case, i.e., saving HTML into YAML files after the HTML
markup has been compiled from RST sources. The API is not
documented and it is quite hard to locate it in the code base and
to understand its purpose, so it could not be used when this
feature was implemented for the first time.