Feedback questionnaire allows non-zero max points (#54)
If none of the questions in a feedback questionnaire has points
set, the questionnaire can still have non-zero max points (that
are set in the questionnaire directive arguments).

The questionnaire max points are checked against the sum of
question points and they are supposed to match, but this commit
loosens the check for feedback questionnaires since they often
award full points for just answering something. Marking points to
each question is then unnecessary.