Fix YAML syntax errors in the rewriting of links (#58)
Previously, the rewriting of links for A+ could introduce syntax errors
in the exercise configuration YAML files. The error was caused by
unescaped quotes in HTML tags and attributes when the code added HTML
data attributes to the content. If the YAML string was delimited by
double quotes ("), then adding data-aplus-path="" into the string
caused an error.

This commit fixes the syntax errors by using the PyYAML library.
The library escapes the YAML content when necessary before it is
written to the disk. The links are still rewritten with regular
expressions as before. In order to rewrite links in the YAML data,
the string fields of the data are iterated one at a time.