Fix links in multilingual courses and exercise descriptions (#58)
Links pointing to chapters in multilingual courses did not work
because the code did not take the ".html" file name ending into
account when removing the language suffix "_en". Sphinx compiles
internal links with relative file paths including the file name
extension. Internal links are produced with, e.g., the `doc` and
`ref` roles.

Links in exercise descriptions pointing to chapters did not work
most of the time, particularly if the chapter was in a nested
directory inside the module directory.

Links in exercise descriptions pointing to static files did not
work especially if the chapter embedding the exercise was in
a nested directory inside the module directory.