Add the show-model option to questionnaires (#60)
The option can be used to disable the model solution of the
questionnaire in A+. Students can not then open the model solution
after the module deadline (with the "show model answer" button
in the exercise navigation bar). By default, the model solution
is shown to students after the module deadline.

The field list setting `questionnaire-default-show-model` in
the index.rst can be used to set a default value for all
questionnaires.

Related to:
https://github.com/Aalto-LeTech/mooc-grader/pull/50