Add postfixes to links so that sphinx finds the files
Courses with language postfixes in their file names such as `first_round_en`
and `first_round_fi` can now leave the postfix from the link. Allthough the
postfix is added in the source read phase of the build, they must be removed
later, since a-plus uses combined chapterkeys.

The recommendable file structure in multilingual courses is to have corresponding
materials in the same folder, using identical names with postfixes. Some courses
use different ways of implementing the structure (for example traky), and this
change wont propably work with them all.