Fix the link paths in the Windows environment
The path to chapters was broken earlier, since splitting the path
in the html_tools.py did not take into account the different
file paths in Windows.

The fix is implemented by replacing '\' with '/', which works fine
in this case, but the operability can not be guaranteed in all
file systems.

The os library did not seem to offer anything usefull, since the folder
names are needed in the nested chapter keys.