Add the js and css code for the annotated directive.
The annotated directive has been part of the a-plus-rst tools for a
long time. However, the annotated directive required a custom
implementation of JS and CSS. This PR add a default JS and CSS
implementation for this directive.

It is important to note that this default implementation can be omitted
by setting the 'include_annotated_js' and 'include_annotated_cs'
parameters to False inside the conf.py file.

Side note: If the 'include_annotated_js' and 'include_annotated_cs' are
set to False or omitted from the conf.py, the user should create and add
their own CSS and JS implementation for the annotated directive.
Otherwise, the directive won't be interactive at all.