Use forward slash in the _rst_srcpath field
Always use the forward slash as the directory separator in the
_rst_srcpath field in exercise YAML files (even when the course is
built on Windows). It is better that the build output is the same
on any platform. This is especially important in courses that
store the RST _build directory in git due to various reasons and
hacks. (For example, Programming 1)