Automatically add language suffixes to doc and ref links
Add the language suffix to doc and ref link targets as well as ref
link labels automatically in multilingual courses.

It is more convenient to write doc links without manually added
language suffixes, e.g., :doc:`chapter1` instead of
:doc:`chapter1_en`. Sphinx can not compile the link if the target
file does not exist, thus the language suffixes must be added
automatically in the first read phase of the compilation.

Likewise, it is convenient to write identical ref link labels in
the same place in all language versions of the chapter. Sphinx
requires that labels are unique, thus language suffixes are
automatically appended to the labels. The ref links in the RST
chapters also refer to the labels without the language suffixes.
The language suffixes are added automatically to the ref links.

If the course uses a different format in links or for some other
reason links need to stay untouched, set
enable_doc_link_multilang_suffix_correction to False in order to
disable doc link modifications and
enable_ref_link_multilang_suffix_correction to False in order to
disable ref link and label modifications. The variables are
defined in conf.py in the course directory (git repository).

Fixes https://github.com/apluslms/a-plus-rst-tools/issues/79

The pull request https://github.com/apluslms/a-plus-rst-tools/pull/65
attempted to change doc links, but it did not work in all cases.