Fix local links in built HTML files
If a course developer wants to open built HTML files directly
under the course _build/html directory, it is important that the
relative links between chapters work locally. The local links
have been broken since the commit
https://github.com/apluslms/a-plus-rst-tools/commit/5ef19aab922a9d5304f7ef70c26a3723346f0894
(pull request https://github.com/apluslms/a-plus-rst-tools/pull/58).
This commit fixes them again. The links built by Sphinx are
modified so that link paths start from the root (_build/html) with
"../". Such links work locally and they also contain the directory
structure so that A+ can fix the links when the chapters are
opened in A+. Knowledge of the directory structure is needed
especially when chapter RST files are nested in subdirectories
under the module directories. Some courses use such subdirectories.

This is the a-plus-rst-tools change related to the issue
https://github.com/apluslms/a-plus/issues/509

Closes https://github.com/apluslms/a-plus-rst-tools/pull/65
The pull request did not work in all cases.