Fix the layout of the next/previous chapter links in the built html
The previous chapter link is now positioned in the left side and
the next chapter link in the right.

This does not affect A+. This affects only the built HTML files
when they are opened from the _build directory, that is,
the "mock frame". A+ does not include these links in the A+ page
content since they are outside the <div id="exercise">.
A+ generates similar links itself.