Disable multilingual link suffix correction by default
Set the following configuration options to False by default:
* `enable_doc_link_multilang_suffix_correction`
* `enable_ref_link_multilang_suffix_correction`

Courses shall enable them only when they really want to use that
feature. Some courses have had issues when these are enabled by default.

In addition, there was a crash during the build in the function
`toc_config.add_lang_suffix_to_links`. It used to double-check that
the course is really multilingual. However, that requires reading
the root toctree, which is not necessarily available during
the source-read phase when the function is called. It is only available
if the index filename is alphabetically before the other files
since Sphinx parses files in the alphabetical order of the filenames.
Fortunately, it is not really necessary to check the root toctree at all
since we can now assume that the options
(enable_{doc,ref}_link_multilang_suffix_correction) are only set when
the course is multilingual and follows the expected file naming scheme.

Fixes #117