Add new autosave option
This commit adds the new `autosave` option to the `questionnaire`
directive that enables autosave for that questionnaire.
In addition, the `enable_autosave` setting in `conf.py` enables
autosave in all questionnaires in the course by default.

When enabled, the `data-aplus-autosave` attribute is inserted
to the exercise HTML. This attribute is used in the A+ frontend
code.

Related to apluslms/a-plus#837