Add exercise container.mounts file paths to the grader files
If exercise config.yaml uses the container.mounts field, then those
file paths must be sent to the grader so that the grader could
really mount those files to the exercise grading container.
This commit adds the file paths to the grader configure settings
so that teachers do not need to manually list these files
in the configure settings.

Fixes #143