Fix the configure files paths in config.yaml container.mounts
The previous commit set the file paths incorrectly for the exercise
files to be configured in the grader. The files are sent from
the Git manager to the grader and thus, the target and source paths
both use paths inside the git repository. The mounting point
targets in the grading container are not part of this setting
even though the previous commit had set those file paths here.

Fixes #143